Refactoring dropbox into operations
