--- Backend/src/main/java/idealab/api/controller/DropboxFile.java
@@ -1,19 +1,0 @@
-package idealab.api.controller;
-
-public class DropboxFile{
-  private String name;
-
-  private DropboxFile(){}
-
-  public DropboxFile(String name){
-    this.name = name;
-  }
-
-  public String getName() {
-    return name;
-  }
-
-  public void setName(String name) {
-    this.name = name;
-  }
-}
